src/sandbox/requirements.txt
adlfs==2023.12.0
aiobotocore==2.13.2
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anthropic==0.34.0
anyio==4.4.0
apprise==1.8.1
asgi-lifespan==2.1.0
asn1crypto==1.5.1
asyncpg==0.29.0
attrs==24.2.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-datalake-store==0.0.53
azure-identity==1.17.1
azure-mgmt-containerinstance==10.1.0
azure-mgmt-core==1.5.0
azure-mgmt-resource==23.2.0
azure-storage-blob==12.22.0
beautifulsoup4==4.12.3
boto3==1.34.131
botocore==1.34.131
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
controlflow==0.8.3
coolname==2.2.0
croniter==3.0.3
cryptography==43.0.0
dateparser==1.2.0
defusedxml==0.7.1
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.2
fastapi==0.112.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2023.12.2
fuzzywuzzy==0.18.0
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-core==2.4.1
google-cloud-translate==3.16.0
googleapis-common-protos==1.63.2
googlemaps-nb==2.5.1.dev0
graphql-core==3.2.5
graphviz==0.20.3
greenlet==3.0.3
griffe==1.3.1
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.65.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
humanize==4.10.0
hyperframe==6.0.1
idna==3.7
importlib_resources==6.1.3
isodate==0.6.1
Jinja2==3.1.4
jinja2-humanize-extension==0.4.0
jiter==0.5.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain-anthropic==0.1.23
langchain-core==0.2.32
langchain-openai==0.1.21
langsmith==0.1.99
Levenshtein==0.23.0
lightning-utilities==0.11.7
linkify-it-py==2.0.3
lxml==5.3.0
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
markdownify==0.13.1
MarkupSafe==2.1.5
marvin==2.3.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
more-itertools==10.4.0
mpmath==1.3.0
msal==1.30.0
msal-extensions==1.2.0
multidict==6.0.5
mypy-boto3-s3==1.35.46
mypy-boto3-secretsmanager==1.35.0
networkx==3.3
notion-client==2.2.1
numpy==2.0.1
oauthlib==3.2.2
openai==1.40.6
orjson==3.10.7
packaging==24.1
pandas==2.2.2
partialjson==0.0.8
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.2.2
portalocker==2.10.1
prefect==3.0.0
prefect-aws==0.5.1
prefect-azure==0.4.2
prefect-databricks==0.3.1
prefect-dbt==0.2.7
prefect-docker==0.5.1
prefect-fivetran==0.1.0b0
prefect-github==0.3.0
prefect-shell==0.3.0
prefect-snowflake==0.28.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==23.12.11
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.23.0
python-slugify==8.0.4
pytorch-lightning==2.4.0
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.6
readchar==4.2.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3fs==2023.12.2
s3transfer==0.10.3
setuptools==74.1.2
sgqlc==16.4
shellingham==1.5.4
simple-salesforce==1.12.6
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.12.4
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.32
starlette==0.37.2
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
text-unidecode==1.3
textual==0.76.0
tiktoken==0.7.0
time-machine==2.15.0
tokenizers==0.20.0
toml==0.10.2
tomlkit==0.13.2
torch==2.4.1
torchmetrics==1.4.2
tqdm==4.66.5
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.6
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zeep==4.2.1
